# Add native UTF8 support for InputText and remove ImWchar buffer