… docs/CHANGELOG.txt
… imstb_textedit.h