@ocornut 2024-09-05
Here are some changes I made:

- I introduced a `stb_textedit_text()` function. legacy `stb_textedit_key()` path can still call `stb_textedit_text()` but imgui calls this directly. Added equivalent ImGuiInputTextState::OnCharPressed().
- tweaked InputTextCalcTextSize() so it doesn't need to call a function for ASCII. it's a potential debug-mode perf regression for large buffers. using similar code as `ImFont::CalcTextSizeA()` and added note that ideally should would be shared (no hurry with that).
- I added a note that STB_TEXTEDIT_GETCHAR() was at this point more of a `STB_TEXTEDIT_GETBYTE()` and verified that it is used accordingly (ascii compares or block copy).
```cpp
// With our UTF-8 use of stb_textedit:
// - STB_TEXTEDIT_GETCHAR is nothing more than a a "GETBYTE". It's only used to compare to ascii or to copy blocks of text so we are fine.
// - One exception is the STB_TEXTEDIT_IS_SPACE feature which would expect a full char in order to handle full-width space such as 0x3000 (see ImCharIsBlankW).
// - ...but we don't use that feature.
```
- If while working on this (before or after) you can come up with any extra test to add to test suite please let me know! even if you don't want to add them yourself, i'll log every idea.
- We can remove the A suffixes later after this is merged and settled.

This needs further works:
- [ ] Couldn't IMSTB_TEXTEDIT_GETPREVCHARINDEX_IMPL() simply use ImTextFindPreviousUtf8Codepoint() ? much less code and likely faster.
- [ ] STB_TEXTEDIT_MOVEWORDLEFT_IMPL, STB_TEXTEDIT_MOVEWORDRIGHT_MAC, STB_TEXTEDIT_MOVEWORDRIGHT_WIN as we don't handle utf-8 codepoints in there?
- [ ] insert mode using `STB_TEXTEDIT_DELETECHARS(str, state->cursor, 1);` is likely broken?
- [ ] the `InputTextReconcileUndoStateAfterUserCallback()` path is sufficiently obscure and rare that I believe we should design some tests for it.
- [ ] rebase on master.