@ocornut 2024-09-05
I think I have fixed all the things except `InputTextReconcileUndoStateAfterUserCallback()`.
Would you mind reviewing my commits?

If for some reason you need to run or modify tests, the modified test is in a public branch: https://github.com/ocornut/imgui_test_engine/tree/features/input_text_7925

About `InputTextReconcileUndoStateAfterUserCallback()`, notice the comment:
```cpp
// Find the shortest single replacement we can make to get the new text from the old text.
// Important: needs to be run before TextW is rewritten with the new characters because calling STB_TEXTEDIT_GETCHAR() at the end.
// FIXME: Ideally we should transition toward (1) making InsertChars()/DeleteChars() update undo-stack (2) discourage (and keep reconcile) or obsolete (and remove reconcile) accessing buffer directly.
```

The part about needing to run before TextW is rewritten is essentially broken now, for the good reason that we were using this buffer as an "old buffer". Solving this the same way would annoyingly requires us to make a copy of the full text before running the callback. aka possibly worse case every frame since the Always callback is technically allowed to modify character...

PLAN A
- We only do a copy and handle reconcile on selected callbacks: Completion, History. Seems limiting.
- We implement what is suggested in the comment aka make InsertChars()/DeleteChars() update undo stack themselves. And we introduce a slightly breaking feature where directly modifying the buffer will break reconcile.
- Having no reconcile can literally break undo stack and crash: #4947, #4949. Actually it will only crash/mess-up with things if string len changed. So we could I suppose track if InsertChars/DeleteChars() have been used. If they have NOT been used and user set `BufDirty = true` AND `BufTextLen` changed, then we could decide to assert.

That seems all possible to implement. The slight annoyance is that it means multiple calls to e.g. InsertChars() would lead to multiple undo steps, so that's a little bit of a regression now. It feels nicer if that's not the case...

PLAN B
- We introduce a new callback data e.g. `BeginDirectBufEdit()`. InsertChars/DeleteChar can call it, and we instruct people modifying text in callback to call this _first_ instead of setting BufDirty=true. There we can do a copy.

PLAN C
- We get lazy and always do a full copy + reconcile.