@alektron 2024-09-05
First of all, I am very glad my PR is helpful even if it ended up requiring more work on your end than I expected.
Unfortunately I will not be able to work on this until at least Monday. I will get back to you in more detail then.
Some quick notes:

> Couldn't IMSTB_TEXTEDIT_GETPREVCHARINDEX_IMPL() simply use ImTextFindPreviousUtf8Codepoint() ? much less code and likely faster.
I was somehow thinking, I only want to handle correctly formed byte sequences when in reality ImTextFindPreviousUtf8Codepoint is not only much simpler, it also behaves better for ill formed sequences.

> insert mode using STB_TEXTEDIT_DELETECHARS(str, state->cursor, 1); is likely broken?
Indeed, I must have missed that.

>  STB_TEXTEDIT_MOVEWORDLEFT_IMPL, STB_TEXTEDIT_MOVEWORDRIGHT_MAC, STB_TEXTEDIT_MOVEWORDRIGHT_WIN as we don't handle utf-8 codepoints in there?
I believe the functions do not technically have to handle that directly since `ImTextFindPreviousUtf8Codepoint` inside `is_word_boundary_from_right/is_word_boundary_from_left` will basically make it work even when the outer loop is  walking through a multibyte sequence byte by byte. It will however do the same work multiple times per multibyteseq which is indeed a bit wasteful.

> If while working on this (before or after) you can come up with any extra test to add to test suite please let me know! even if you don't want to add them yourself, i'll log every idea.
I'll see what I can do

For everything else I will have to get back to you next week, sorry.
Especially the whole `InputTextReconcileUndoStateAfterUserCallback()` topic. I had some trouble understanding the purpose what exactly it does when working this. I'll have to read into this in more detail first.